requirements.txt
numpy==1.18.1
torch==1.3.1
psutil==5.6.7
pandas==1.0.1
scipy==1.4.1
kombu==4.6.8
joblib==0.14.1
tqdm==4.42.1
xgboost==0.90
chainer==6.3.0
torchvision==0.4.2
Bio==0.1.0
librosa==0.7.2
Pillow==7.1.2
scikit_learn==0.22.2.post1
